pyproject.toml
[project]
name = "fuelcounter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cv2-enumerate-cameras>=1.3.3",
    "flask>=3.1.3",
    "mjpeg-streamer>=2024.2.8",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "scikit-image>=0.26.0",
    "scipy>=1.17.1",
]
